# swagger/spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "hash-challenge-microservice",
    "description": "API HTTP (JSON) de e-commerce (venda online)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:{port}",
      "description": "Local server",
      "variables": {
        "port": {
          "default": "8082"
        }
      }
    }
  ],
  "tags": [
    {
      "name": "API",
      "description": ""
    },
    {
      "name": "Health Check",
      "description": ""
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Rota para monitoramento e checagem de disponibilidade.",
        "tags": [
          "Health Check"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthCheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/checkout": {
      "post": {
        "summary": "Consolida o carrinho de compras, efetua o cáculculo de descontos por produto e adiciona brinde caso elegível.",
        "tags": [
          "API"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSuccess"
                }
              }
            }
          },
          "422":{
            "$ref":"#/components/responses/EntityError"
          },
          "400":{
            "$ref":"#/components/responses/BadRequestError"
          },
          "500":{
            "$ref":"#/components/responses/ServiceError"
          }
        },
        "operationId": "checkout",
        "parameters": [],
        "description": "Consolida o carrinho de compras aplica descontos por produto e inclui brinde caso elegível",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "products": {
                    "type": "array",
                    "description": "produtos",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": "4",
                          "description": "ID do produto"
                        },
                        "quantity": {
                          "type": "number",
                          "example": "2",
                          "description": "quantidade referente"
                        }
                      }
                    }
                  }
                },
                "required": [
                  "products"
                ]
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "something is wrong"
          }
        }
      },
      "BadRequestError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "product(s) out off stock"
          },
          "outOfStock": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": "1"
                },
                "quantity": {
                  "type": "number",
                  "example": "1"
                }
              }
            }
          }
        }
      },
      "CheckoutSuccess": {
        "type": "object",
        "properties": {
          "total_amount": {
            "type": "number",
            "example": "30314",
            "description": "valor total da compra"
          },
          "total_amount_with_discount": {
            "type": "number",
            "example": "30314",
            "description": "valor total da compra com descontos"
          },
          "total_discount": {
            "type": "number",
            "example": "0",
            "description": "valor total dos descontos"
          },
          "products": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1,
                  "description": "ID do produto"
                },
                "quantity": {
                  "type": "number",
                  "example": 2,
                  "description": "quantidade referente"
                },
                "unit_amount": {
                  "type": "number",
                  "example": 15157,
                  "description": "valor unitário"
                },
                "total_amount": {
                  "type": "number",
                  "example": 30314,
                  "description": "valor total da soma dos itens"
                },
                "discount": {
                  "type": "number",
                  "example": 0,
                  "description": "valor total dos descontos"
                },
                "is_gift": {
                  "type": "boolean",
                  "example": false,
                  "description": "flag de brinde"
                }
              }
            }
          }
        }
      },
      "CheckoutResponseOnEntityError": {
        "type":"object",
        "properties":{
          "message": {
            "type":"string",
            "example": "Invalid payload. Get instructions on /docs route"
          }
        }
      },
      "HealthCheckResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Service Up"
          }
        }
      }

    },
    "parameters": {},
    "responses": {
      "EntityError":{
        "description":"Erro de requisição com dados faltantes ou inválidos",
        "content":{
          "application/json":{
            "schema":{
              "$ref":"#/components/schemas/CheckoutResponseOnEntityError"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceError": {
        "description": "Erro de serviço",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequestError": {
        "description": "Erro de requisição mal formada",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BadRequestError"
            }
          }
        }
      }
    }
  }
}